debugging

+++ reifenmodelle/src/main/java/net/reifenapp/reifenmodelle/resource/ReifenmodellController.java
@@ -2,6 +2,9 @@ package net.reifenapp.reifenmodelle.resource;
 
 import net.reifenapp.reifenmodelle.model.Reifenmodell;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import net.reifenapp.reifenmodelle.service.ReifenmodellService;
 import org.springframework.http.HttpStatus;
@@ -31,19 +34,34 @@ public class ReifenmodellController {
 
     @PostMapping("/add")
     public ResponseEntity<Reifenmodell> addReifenmodell(@RequestBody Reifenmodell Reifenmodell) {
-        Reifenmodell addReifenmodell = reifenmodellService.addReifenmodell(Reifenmodell);
-        return new ResponseEntity<>(addReifenmodell, HttpStatus.CREATED);
+        if(checkAuthorization()) {
+            Reifenmodell addReifenmodell = reifenmodellService.addReifenmodell(Reifenmodell);
+            return new ResponseEntity<>(addReifenmodell, HttpStatus.CREATED);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @PutMapping("/update")
     public ResponseEntity<Reifenmodell> updateReifenmodell(@RequestBody Reifenmodell Reifenmodell) {
-        Reifenmodell reifenmodell = reifenmodellService.updateReifenmodell(Reifenmodell);
-        return new ResponseEntity<>(reifenmodell, HttpStatus.OK);
+        if(checkAuthorization()) {
+            Reifenmodell reifenmodell = reifenmodellService.updateReifenmodell(Reifenmodell);
+            return new ResponseEntity<>(reifenmodell, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<?> updateReifenmodell(@PathVariable("id") Long id) {
-        reifenmodellService.deleteReifenmodell(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+        if(checkAuthorization()) {
+            reifenmodellService.deleteReifenmodell(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+    }
+
+    //Better if defined in another class (a Service)
+    private boolean checkAuthorization() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        return authentication.getAuthorities().contains(new SimpleGrantedAuthority("admin"));
     }
 }
